change type parameter to avoid warning

+++ LAB2/Radioactive_Contamination_Concentration/Sequence/main/main.h
@@ -1,14 +1,13 @@
-#include <math.h>   // For pow
 const char *in_path = "../../Input/radioactive_matrix.csv";
 const char *out_path = "../../Output/out_radioactive_matrix.csv";
 const int rows = 4000;
 const int cols = 4000;
 const int iterations = 100;
 // One point represent 100 meters
-const int dx = 100; 
-const int dy = 100;
+const float dx = 100; 
+const float dy = 100;
 //  Time step : 1 second
-const int dt = 1;
+const float dt = 1;
 // Wind velocity    
 const float ux = 3.3;
 const float uy = 1.4;
